requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
pycocotools>=2.0.6
tqdm>=4.65.0
albumentations>=1.3.0
gradio>=4.0.0
requests>=2.31.0
